refactor: move summary form to component
